requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
google-generativeai==0.3.2
APScheduler==3.10.4
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0